Refactor pageparts into namespaced class
